% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFbounddata.R
\name{AFbounddata}
\alias{AFbounddata}
\title{Calculate the assumption free bound for a data set.}
\usage{
AFbounddata(outcome, treatment, selection, whichPar)
}
\arguments{
\item{outcome}{A vector with the binary outcome variable.}

\item{treatment}{A vector with the binary treatment variable.}

\item{selection}{A vector with the binary selection variable or a scalar representing the probability of being selected.}

\item{whichPar}{A string defining the population parameter of interest. Available options are as follows. (1) Relative risk in the total population: "RR_tot", (2) Risk difference in the total population: "RD_tot", (3) Relative risk in the subpopulation: "RR_s", (4) Risk difference in the subpopulation: "RD_s".}
}
\value{
A scalar with the assumption free bound.
}
\description{
Calculate the assumption free bound for a data set.
}
\examples{
y = c(0,0,0,0,1,1,1,1)
tr = c(0,0,1,1,0,0,1,1)
sel = c(0,1,0,1,0,1,0,1)
AFbounddata(y,tr,sel,"RR_tot")

y = c(0,0,0,0,1,1,1,1)
tr = c(0,0,1,1,0,0,1,1)
selProb = 0.5
AFbounddata(y,tr,selProb,"RR_tot")
}
