% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVboundM.R
\name{SVboundM}
\alias{SVboundM}
\title{Calculate the Smith and VanderWeele bound for the M structure (Smith, L. H., & VanderWeele, T. J. (2019). Bounding bias due to selection.).}
\usage{
SVboundM(pVprob, pUprob, pTcoef, pY1coef, pY0coef, pScoef, whichPar, probit)
}
\arguments{
\item{pVprob}{A scalar that represents the probability that V=1.}

\item{pUprob}{A scalar that represents the probability that U=1.}

\item{pTcoef}{A numerical vector with two elements. The first element is the intercept in the model for the treatment (as the defined by the input "probit"). The second element is the slope in the model for the treatment (as defined by the input "probit").}

\item{pY1coef}{A numerical vector with two elements. The first element is the intercept in the model for the potential outcome Y(1) (as the defined by the input "probit"). The second element is the slope in the model for the potential outcome Y(1) (as defined by the input "probit").}

\item{pY0coef}{A numerical vector with two elements. The first element is the intercept in the model for the potential outcome Y(0) (as the defined by the input "probit"). The second element is the slope in the model for the potential outcome Y(0) (as defined by the input "probit").}

\item{pScoef}{A matrix of size K by 4, where K is the number of selection variables. The first column is the intercepts in the models for the selection variables (as defined by the input "probit"). The second column is the slopes for V in the models for the selection variables (as defined by the input "probit"). The third column is the slopes for U in the models for the selection variables (as defined by the input "probit"). The fourth column is the slopes for T in the models for the selection variables (as defined by the input "probit").}

\item{whichPar}{A string defining the population parameter of interest. Available options are as follows. (1) Relative risk in the total population: "RR_tot", (2) Risk difference in the total population: "RD_tot", (3) Relative risk in the subpopulation: "RR_s", (4) Risk difference in the subpopulation: "RD_s".}

\item{probit}{A logical variable that is used to define the models for the variables in the M structure. If T, the probit model is used. If F, the logit model is used.}
}
\value{
A list containing the Smith and VanderWeele bound and an indicator if the treatment has been reversed.
}
\description{
Calculate the Smith and VanderWeele bound for the M structure (Smith, L. H., & VanderWeele, T. J. (2019). Bounding bias due to selection.).
}
\examples{
pV = 0.1
pU = 0.1
pT = c(0,1)
pY1 = c(0,1)
pY0 = c(0,1)
pS = matrix(c(1,0,0,0,1,0,0,0),nrow=2,byrow=TRUE)
SVboundM(pV,pU,pT,pY1,pY0,pS,"RR_tot",TRUE)
}
