% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVmarginalM.R
\name{SVmarginalM}
\alias{SVmarginalM}
\title{Calculate the Smith and VanderWeele bound for the M structure, for a margianl causal estimand (Smith, L. H., & VanderWeele, T. J. (2019). Bounding bias due to selection.).}
\usage{
SVmarginalM(a, b)
}
\arguments{
\item{a}{CHANGE}

\item{b}{CHANGE}
}
\value{
CHANGE
}
\description{
Calculate the Smith and VanderWeele bound for the M structure, for a margianl causal estimand (Smith, L. H., & VanderWeele, T. J. (2019). Bounding bias due to selection.).
}
\examples{
CHANGE
}
