% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFbound.R
\name{AFbound}
\alias{AFbound}
\title{Calculate the assumption free bound for a data set}
\usage{
AFbound(outcome, treatment, selection, whichEst)
}
\arguments{
\item{outcome}{Input vector. A binary outcome variable.}

\item{treatment}{Input vector. A binary treatment variable.}

\item{selection}{Input vector or input scalar. A binary selection variable or
a selection probability.}

\item{whichEst}{Input string. Defining the population parameter of interest.
Available options are as follows. (1) Relative risk in the total
population: "RR_tot", (2) Risk difference in the total population:
"RD_tot", (3) Relative risk in the subpopulation: "RR_s", (4) Risk
difference in the subpopulation: "RD_s".}
}
\value{
A scalar with the assumption free bound.
}
\description{
\code{AFbound()} returns the assumption free bound for an observed data set
that consists of an outcome, a treatment and a selection variable.
}
\examples{
y = c(0,0,0,0,1,1,1,1)
tr = c(0,0,1,1,0,0,1,1)
sel = c(0,1,0,1,0,1,0,1)
AFbound(y,tr,sel,"RR_tot")

y = c(0,0,0,0,1,1,1,1)
tr = c(0,0,1,1,0,0,1,1)
selProb = 0.5
AFbound(y,tr,selProb,"RR_tot")
}
