% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVboundsharp.R
\name{SVboundsharp}
\alias{SVboundsharp}
\title{Check if the Smith and VanderWeele bound in the subpopulation is sharp}
\usage{
SVboundsharp(BF_U, prob, SVbound = NULL, AFbound = NULL)
}
\arguments{
\item{BF_U}{Input scalar. The bounding factor for the SV bounds in the
subpopulation. Must be equal to or above 1.}

\item{prob}{Input scalar. The probability P(Y=1|T=0,I_S=1). The eventual
recoding has to be done manually.}

\item{SVbound}{Optional input scalar. The SV bound (calculated in another
place). Only necessary if one wants to know if the SV bound is not sharp.}

\item{AFbound}{Optional input scalar. The AF bound (calculated in another
place). Only necessary if one wants to know if the SV bound is not sharp.}
}
\value{
A string stating if the SV bound is sharp, or if it is inconclusive.
}
\description{
\code{SVboundsharp()} returns a string that indicates if the SV bound is sharp, if
it's inconclusive or if it's not sharp.
}
\examples{
SVboundsharp(BF_U = 2, prob = 0.4, SVbound = 2, AFbound = 3)
SVboundsharp(BF_U = 2, prob = 0.9, SVbound = 2, AFbound = 1.8)

}
\references{
Smith, Louisa H., and Tyler J. VanderWeele. "Bounding bias due to
selection." Epidemiology (Cambridge, Mass.) 30.4 (2019): 509.

Zetterstrom, Stina and Waernbaum, Ingeborg. MANUSCRIPT XXX
}
